Move Markdown function out of helpers into markdown-extra
