+++ src/ProofFlow/plugins/plugin-markdown.ts
@@ -4,15 +4,13 @@ import { TextSelection } from "prosemirror-state";
 
 import {
   isClickedNode,
-  renderedToMarkdown,
-  markdownToRendered,
   highLevelCells,
   getContainingNode,
 } from "../commands/helpers.ts";
 import { ProofFlowSchema } from "../editor/proofFlowSchema.ts";
 import { proofFlow } from "../../main.ts";
 import { UserMode, lockEditing } from "../UserMode/userMode.ts";
-import { markdownRenderedClickFix } from "./markdown-extra.ts";
+import { markdownRenderedClickFix, markdownToRendered, renderedToMarkdown } from "./markdown-extra.ts";
 
 /**
  * Plugin that handles the conversion between markdown and rendered markdown nodes.
